Move `free_port` related fixtures
